requirements.txt
tqdm
scipy
numpy
matplotlib
torch
torchvision
pillow
datasets
pandas
pyarrow